extract data

+++ helpers.py
@@ -1,7 +1,10 @@
-import os, shutil
+import os
+import shutil
+import json
 import requests
 import sys
 from zipfile import ZipFile
+import pandas as pd
 
 import bcolor
 
@@ -28,7 +31,7 @@ def download(url: str, save_path: str = "data", file_name: str = "data.zip"):
     
     try:
         with open(path_file, "wb") as f:
-            res = requests.get(url, stream=True)
+            res = requests.get(url, stream=True, timeout=5)
             total_length = int(res.headers.get('content-length', 0))
             
             if total_length is None: # no content length header
@@ -44,6 +47,16 @@ def download(url: str, save_path: str = "data", file_name: str = "data.zip"):
             f.close()
 
         return path_file
+    
+    except requests.HTTPError as e:
+        print("\nChecking internet connection failed, status code {0}.".format(
+                e.response.status_code
+            )
+        )
+        return None
+    except requests.ConnectionError:
+        print("\nNo internet connection available.")
+        return None
     except:
         return None
     
@@ -65,6 +78,8 @@ def unzip(zip_file: str, save_folder: str = "./data", new_file: str = "data.json
         if os.path.exists(zip_file):
             print("Rm file " + zip_file)
             os.remove(zip_file)
+        
+        return new_file
     except:
         return None
     
@@ -77,9 +92,9 @@ def rm_all_file_in_folder(folder: str):
     # rm all file in folder
     for filename in os.listdir(folder):
         file_path = os.path.join(folder, filename)
-        
+
         try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
+            if os.path.isfile(file_path) or os.path.islink(file_path) and file_path != ".gitignore":
                 os.unlink(file_path)
             elif os.path.isdir(file_path):
                 shutil.rmtree(file_path)
@@ -88,3 +103,32 @@ def rm_all_file_in_folder(folder: str):
         except Exception as e:
             print('Failed to delete %s. Reason: %s' % (file_path, e))
             return False
+        
+def load_file_json(path_file=''):
+    try:
+        with open(path_file, 'r') as f:
+            content = json.load(f)
+        return content
+    except:
+        return None
+    
+def write_file_json(var_dict: dict, path_file: str =' '):
+    if var_dict:
+        try:
+            with open(path_file, 'w') as outfile:
+                json.dump(var_dict, outfile, indent=4, ensure_ascii=False)
+
+            return True
+        except:
+            return False
+
+    return False
+
+def json_to_excel(data, save_file):
+    try:
+        df = pd.json_normalize(data)
+        df.to_excel(save_file)
+        
+        return True
+    except:
+        return False
